feat(peer): add thread-safe peer table
